+++ bukkit/src/main/java/me/u8092/mirlo/bukkit/commands/ReloadCommand.java
@@ -1,7 +1,7 @@
-package me.u8092.watchdog.bukkit.commands;
+package me.u8092.mirlo.bukkit.commands;
 
-import me.u8092.watchdog.bukkit.Main;
-import me.u8092.watchdog.bukkit.util.DebugUtil;
+import me.u8092.mirlo.bukkit.util.DebugUtil;
+import me.u8092.mirlo.bukkit.Main;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -9,7 +9,7 @@ import org.bukkit.command.CommandSender;
 public class ReloadCommand implements CommandExecutor {
     @Override
     public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-        if(!sender.hasPermission("watchdog.reload")) return false;
+        if(!sender.hasPermission("mirlo.reload")) return false;
 
         Main.getInstance().reloadConfig();
         if(Main.getInstance().getConfig().getBoolean("debug")) DebugUtil.info("config.yml has been reloaded");
