+++ bukkit/src/main/java/me/u8092/mirlo/bukkit/util/DebugUtil.java
@@ -1,6 +1,6 @@
-package me.u8092.watchdog.bukkit.util;
+package me.u8092.mirlo.bukkit.util;
 
-import me.u8092.watchdog.bukkit.Main;
+import me.u8092.mirlo.bukkit.Main;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.Bukkit;
@@ -10,20 +10,20 @@ public class DebugUtil {
     public static void info(String message) {
         Main.getInstance().getLogger().info(message);
         String PREFIX = Main.getInstance().getConfig().getString("debug-prefix");
-        if(PREFIX == null) PREFIX = "[WatchdogDebug]";
+        if(PREFIX == null) PREFIX = "[MirloDebug]";
 
         if(Main.isPaper()) {
             Component playerMessage = Component.text(PREFIX + " ", NamedTextColor.DARK_GREEN)
                     .append(Component.text(message, NamedTextColor.GRAY));
 
             for(Player player : Bukkit.getOnlinePlayers()) {
-                if(player.hasPermission("watchdog.debug.receive")) player.sendMessage(playerMessage);
+                if(player.hasPermission("mirlo.debug.receive")) player.sendMessage(playerMessage);
             }
             return;
         }
 
         for(Player player : Bukkit.getOnlinePlayers()) {
-            if(player.hasPermission("watchdog.debug.receive")) player.sendMessage(PREFIX + " " + message);
+            if(player.hasPermission("mirlo.debug.receive")) player.sendMessage(PREFIX + " " + message);
         }
     }
 }
