+++ commons/pom.xml
@@ -5,11 +5,11 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>me.u8092</groupId>
-        <artifactId>watchdog</artifactId>
-        <version>0.2.0d</version>
+        <artifactId>mirlo</artifactId>
+        <version>0.2.0e</version>
     </parent>
 
-    <artifactId>watchdog-commons</artifactId>
+    <artifactId>mirlo-commons</artifactId>
     <version>1.0.0</version>
 
     <properties>
@@ -17,5 +17,4 @@
         <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
-
 </project>
